+++ DelvUI/changelog.md
@@ -1,5 +1,6 @@
 # 1.7.1.0
 - Added "Different Size when targeted" setting for Nameplates.
+- Added "Custom Color when being targeted" setting for Enemy Nameplates.
 - Fixed "Full" Occlusion Mode for Nameplates.
 
 # 1.7.0.0